WIP: refactoring, ln_p2p
